4639 :Fix for  LX Accessibility bug related to filter.
[LX][Search results - Sort and filter - Things to do][TalkBack] Focus is lost to the top of the screen upon toggling checkboxes

https://eiwork.mingle.thoughtworks.com/projects/local_expert/cards/4639

Screen reader users will have to navigate back to the list of filter checkboxes from the top of the screen after selecting any of filter.
Additionally, screen reader users who are blind may be disoriented by the unexpected transition.
When user selects any filter, the filters are deleted and added back into the view which is causing the issue.

Fix is applied to prevent deletion and adding back when filter is selected.