4286 : Fix for LX Accessibility Bug related to search icon
[LX][Infosite - Things to do][TalkBack] The header's search button is at the end of the screen's swipe order.
https://jira.sea.corp.expecn.com/jira/browse/WCAG-4286
The back (X-navigation icon ) and search icon are aadded in runtime to the toolbar view.
The talkback is not able to add the elements appearance order in the accessibility order chain.
Programmatically the focus is applied on navigation icon as the first element in the accessibility chain in the existing code.