Use Java reflection for remote flags (#10532)
Kotlin reflection is proving too hard to get working in the face of proguard.
Without proguard I get a property that returns me the `Feature`. With proguard
I get a property that returns me the `RemoteFeatureDelegate`. It's really
strange and unintuitive.

The java reflection is much more straight forward. Find all the methods that
return `Feature`.